test(resolver): split and harden plugin resolver specs
