+++ cli/src/test/kotlin/me/liam/microsmith/cli/plugins/PluginResolverGuardrailTests.kt
@@ -5,9 +5,7 @@ import io.kotest.matchers.collections.shouldBeEmpty
 import io.kotest.matchers.shouldBe
 import io.kotest.matchers.types.shouldBeTypeOf
 import me.liam.microsmith.cli.command.RunCommand
-import java.nio.file.Path
 import kotlin.io.path.ExperimentalPathApi
-import kotlin.io.path.createTempDirectory
 import kotlin.io.path.deleteRecursively
 import kotlin.io.path.writeBytes
 import kotlin.io.path.writeText
@@ -16,7 +14,7 @@ import kotlin.io.path.writeText
 class PluginResolverGuardrailTests :
     StringSpec({
         "returns success without evaluating credential resolvers when no plugins are requested" {
-            val tempDir = createTempDirectory("microsmith-plugin-resolver-no-plugins")
+            val tempDir = createWorkingDirectoryTempDirectory("microsmith-plugin-resolver-no-plugins")
             try {
                 val command =
                     RunCommand(
@@ -48,7 +46,7 @@ class PluginResolverGuardrailTests :
         }
 
         "resolves local plugin jars without evaluating repository credential resolvers" {
-            val tempDir = createTempDirectory("microsmith-plugin-resolver-local-only")
+            val tempDir = createWorkingDirectoryTempDirectory("microsmith-plugin-resolver-local-only")
             try {
                 val script = tempDir.resolve("schema.microsmith.kts")
                 script.writeText("// test script")
@@ -56,8 +54,7 @@ class PluginResolverGuardrailTests :
                 localJar.parent?.toFile()?.mkdirs()
                 localJar.writeBytes("local-plugin".toByteArray())
 
-                val workingDirectory = Path.of("").toAbsolutePath().normalize()
-                val relativeJarPath = workingDirectory.relativize(localJar.toAbsolutePath().normalize())
+                val relativeJarPath = relativizeFromWorkingDirectory(localJar)
                 val command =
                     RunCommand(
                         script = script,
@@ -82,6 +79,40 @@ class PluginResolverGuardrailTests :
                 val result = resolvePlugins(command = command, settings = settings)
                 val success = result.shouldBeTypeOf<PluginResolutionResult.Success>()
                 success.classpath shouldBe listOf(localJar.toAbsolutePath().normalize())
+            } finally {
+                runCatching { tempDir.deleteRecursively() }
+            }
+        }
+
+        "returns failure diagnostics when credential sensitive values cannot be initialized" {
+            val tempDir = createWorkingDirectoryTempDirectory("microsmith-plugin-resolver-credential-init")
+            try {
+                val command =
+                    RunCommand(
+                        script = tempDir.resolve("schema.microsmith.kts"),
+                        outputDir = tempDir.resolve("generated"),
+                        plugins = setOf("com.acme:plugin:1.0.0"),
+                    )
+                val settings =
+                    PluginResolverSettings(
+                        cacheDirectory = tempDir.resolve("cache"),
+                        repositoryCredentialsResolver =
+                        object : RepositoryCredentialsResolver {
+                            override fun resolve(repositoryUri: String): RepositoryCredentials? {
+                                error("resolve() should not be called when sensitive value initialization fails.")
+                            }
+
+                            override fun sensitiveValues(): Set<String> {
+                                throw IllegalArgumentException("Repository credentials file is invalid.")
+                            }
+                        },
+                    )
+
+                val result = resolvePlugins(command = command, settings = settings)
+                val failure = result.shouldBeTypeOf<PluginResolutionResult.Failure>()
+                failure.diagnostics.joinToString("\n").shouldBe(
+                    "[authentication] Repository credentials file is invalid.",
+                )
             } finally {
                 runCatching { tempDir.deleteRecursively() }
             }
